PHP client: library files to a separate directory
structure suitable for PSR-4 autoloading
