+++ basex-api/src/main/php/load.php
@@ -1,0 +1,12 @@
+<?php
+/*
+ * This script loads all code of the PHP BaseXClient library for use in the example scripts.
+ * Usually the library won't be used like this, but copied to a project and loaded by PSR-4 autoload.
+ *
+ * Documentation: https://docs.basex.org/wiki/Clients
+ *
+ * (C) BaseX Team 2005-12, BSD License
+ */
+
+include_once 'BaseXClient/Session.php';
+include_once 'BaseXClient/Query.php';
